Removed min max arrays for separate fields and removed Fire method
Instead of using arrays of length 2 to store min max values, which
is very stupid, now two variables are used. While on it, I decided
it'd be better to extend Random with a method that accepts floats
and instantiate particles in-place rather than having a Fire method
within ParticleGenerator just to create a particle.