Restructured scoring events
This is a big one.
Along with fixing set win conditions (which were wrong), the
idea here was to make scoring events more intuitive. This is
how it works now:

Whenever a point is scored, the Score event will be invoked, and,
a function subscribed to that event do the following:

* Make the match not in-play

* Check if the scorer was on game/set/match point and invoke the
correspoding event. A function subscribed to that event will
handle adding score, stamina and service.

* Setup the transition and start it (part of which is handed off
to the UI manager)

Additionally, I deemed the transition label to be kind of hacky
so I removed it. I'll look to implement something similar later.