Fix double pressing having unintended behaviour
Holding any input should stop the buffer from counting towards clear,
which allows motion inputs like DP to be done while walking forward,
that is to say, while the first input was being held. The problem with
this is that behaviour would feel weird with sprinting, so originally
we just made sure heldTime of the first press wasn't over ClearTime,
meaning the first input towards a sprint wasn't being held beforehand.
However, this wasn't possible because the heldTime of an input is
increased if its action is being held, meaning that there was no way
to disern if the first input was held beforehand or it was just that
the second input was held. The fix to this was to only count heldTime
for the last input in the buffer but that brought problems too, so the
final solution to this mess was to, in the case of having multiple
inputs with the same action within the buffer, only count te heldTime
of the last one of those which share the same action.