Simplify and improve InputManager
Now the clearing strategy is even simpler: Clear all non-held inputs and
held inputs are only spared if they have the same action the last input
has. This compromise is perfect, it preserves both the ability to hold
the first input of a motion input beforehand and the ability to
sprint-jump while also preventing players from starting a sprint after
a another sprint with just one input, instead of the more intuitive two.